meiduo_mall/requirements.txt
amqp==5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.0.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.11.0
django-crontab==0.7.1
django-filter==2.4.0
django-haystack==3.0
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
elasticsearch==2.4.1
fdfs-client==4.0.7
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.2.3
MarkupSafe==1.1.1
mutagen==1.45.1
packaging==21.3
Pillow==8.2.0
prompt-toolkit==3.0.28
pycparser==2.21
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.7
python-alipay-sdk==3.0.4
pytz==2021.3
QQLoginTool==0.3.0
redis==4.1.4
requests==2.27.1
ronglian-sms-sdk==1.0.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
